requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
numpy==2.4.2
pandas==3.0.0
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
Werkzeug==3.1.5
